requirements.txt
matplotlib==1.3.1
scipy==0.19.0
tensorflow-gpu==1.1.0
numpy==1.11.1
Pillow==4.1.1